Updates base styles to use `:where()`.
`:where()` has *zero* specificity, so it is very useful for defaults. This allows me to remove the awkward `--dwac-*` variables (that aren't `--dwac-theme-*`), which were necessary to override defaults in certain circumstances. For example, `.frag-footer { color: blue }` would sometimes *not* override a default `a { color: red; }`. However, it will always override `:where(a) { color: red; }`.

This also brought to light a typo'd usage of `--dwac-theme-font-size-medium` and also that the value was off from the real one being used due to defaults in `base.css`. Fixed the variable name and updated its value to be more useful.