Adds `markdown-it-anchor` to add anchor IDs to section headers.
This adds support for `#some-header` to the URL bar so it jumps to the relevant section.

I also added `slugify` to perform the conversion to a string, which provides better handle for edge cases and unicode. Unfortunately, I needed to use an older version because the newest one is ESM-only.