Adds clickable link icon to headers which copy to clipboard.
This makes headers show a "link" icon on mouseover and when clicked the hash of that header is added to the current window location and the full URL is copied to clipboard. Currently this copy operation is not communicated to the user, but it happens under the hood.

I also removed `line-height` from all the header styles because this was messing up the `focus-visible` box, making it taller where text was present, and shorter where there was empty space. Turns out `line-height: normal;` (the default) just works the way I want and no overriding needs to be done.

This implemented via a `handleHeaderLinkOnClick()` function to facilitate testing. Tests can run with any given node, while the actual page simply has a top-level call to the same function on the `document`.