Fixes Karma `tsconfig.json` to include all client sources.
Apparently this needs to match the Karma config's `files` option to serve all files in dev. Even though test files should depend on their production code, Karma doesn't seem to understand this and needs all of them to be built. This is a bit unfortunate, and might cause issues for files with complex side effects, but it seems to be good enough for now.