Adds `dwac-snackbar` custom element and shows one when a URL is copied to clipboard when clicking on a header.
The snackbar element shows a small bubble fixed at the top of the screen with a short message for the user. It is intended as a non-blocking alert message. I tried to be as UX-friendly with this as possible:

* It uses `role="alert"` which I think is the most applicable for accessibility.
* The custom element actually spans the entire width of the body, but uses `pointer-events: none;` to allow clicks to pass through it as is expected. Only the actual message itself should block clicks.
* The snackbar has a strong black background with solid white text to keep high contrast and improve readability.
* The snackbar fades in and out to make its movement a little less jarring.

The custom element's API is not fully ideal, so I made a wrapper `snackbar.show()` function which creates the element, appends it to the DOM, waits the appropriate time, and then removes it. This is the ideal interface for a simple UI message and hides much of the complexity around the custom element.

This is currently used to inform the user when they click a header that it was copied to their clipboard.

I considered adding an `X` button to close the snackbar explicitly, but decided against it. This was mainly because I'm not totally sure how a11y would work (which is probably solvable) and line wrapping of long messages could get tricky to layout. For such short and ephemeral messages, I don't think an explicit close button is really necessary right now.

I also punted on supporting showing multiple snackbar messages at the same time. For now, they just overlay each other, which is fine since there is only one message which can be displayed. In the future it may be necessary to queue up the messages or offset them in a list.