Fixes snackbar animation in prod builds.
The animation was broken because [`minify-html-literals` chokes on template literal values](https://github.com/asyncLiz/minify-html-literals/issues/31). As a result, we can't inline the animation durations in the template literal. As a workaround, a CSS property is used with the value set in `connectedCallback()`, so the CSS template is a pure literal with no values.

Arguably these properties should be set on the `:root { }` node exactly once, rather than on each `<dwac-snackbar />` element, but that would be harder to test and the performance benefits are likely negligible.