Adds mime type to `<picture />` source elements.
The `type` attribute isn't strictly required, but when it isn't present the browser must make a network request for each source until it finds one that the browser can render, which is an unnecessary performance hit. Instead, we generate the `type` attribute based on the source's file extension, which should save the browser a couple round trips.

This also includes one minor fix to trim alt text to remove any leading or trailing whitespace.