Fixes share component when not browser does not support sharing.
On Firefox (which doesn't support `navigator.share`), this component was displaying "false" instead of the share icon. This is because I misunderstood the intended use of [`ifDefined()`](https://lit.dev/docs/api/directives/#ifDefined), which is supposed to set attributes, not conditionally render DOM elements.

[`when()`](https://lit.dev/docs/api/directives/#when) is the API that does what I originally wanted. Switched over to use that which now renders correctly in Firefox.

I added a test to confirm that the share button isn't shown when not supported, though that technically wouldn't have caught this bug anyways since the element wasn't rendered, it just displayed "false" in its place.