Inlines the logo with header text.
Previously, if the viewport width was too narrow, the header would line break between the logo and "Devel", which doesn't look very good. Instead, we move the logo inside the `<h1 />` tag so it renders properly inline and word wraps correctly.

Needed to adjust the vertical align slightly, as `vertical-align: middle;` is simply not accurate. Workaround was to use `display: relative; top: _px;` to adjust it to the right spot. Hopefully this is good enough across browers?