Adds image to Twitter metadata.
I left out image because they explicitly recommended *not* to put a generic image of the author or site, but their substitute is an even more generic "article" icon, which I don't see as an improvement. I believe this should be better.

Twitter does have some requirements for the image, notably that it must be between 144x144 & 4096x4096 and < 5MB in size. I presume it also needs a 1:1 aspect ratio, but their wording is confusing. The image as is meets the requirements, though it might be good to make a separate URL in the future in case the image changes and breaks one of the requirements.