Adds base script.
This updates the `base.njk` template to allow scripts to be passed in by URL. Every page using the base template uses the new `/scripts/base.js` file (which is currently empty), while posts continue to use `/scripts/post.js`, which now composes the base script.

This allows a singular hook of JavaScript which will run for every page on the site.

Unfortunately every page needs to explicitly link in the base script, as there's no way to do this by default in a overridable fashion. I could have set it up so the post page loads `base.js` and `post.js` in two `<script />` tags, but doing so would duplicate shared dependencies between them. This does make things less cacheable, since content downloaded from `base.js` needs to be redownloaded for `post.js`, but I think this is a minor concern.