Refactors `::: timestamp :::` markdown plugin into a Nunjucks expression.
Apparently markdown files are preprocessed by the templating engine before being parsed as markdown. This used Liquid as the default engine but was unused.

This commit changes the templating language to Nunjucks to match the rest of the codebase. It also adds a new `postDate` filter which is used to render the timestamp into the correct HTML.