Revert "Refactors `::: timestamp :::` markdown plugin into a Nunjucks expression."
This reverts commit e955d1e585d9c7027857fc98023279abfaf8be58.

Turns out, we don't want to use Nunjucks expressions in a markdown file because it couples the file to both of these templating system. In particular, it means that moving to a custom markdown implementation (`marked`) is harder because this doesn't work out of the box. Instead, it's better to just stick with markdown as the source language and write custom plugins/extensions where appropriate rather than relying on Nunjucks on top of markdown.