Adds frontmatter context.
This creates a "context-like" object which is attached to the current stack by `useContext()` and `getContext()`. It allows the top of the stack frame to set some data which is read by the bottom of the stack frame without passing it through all the frames in the middle.

This will serve as a hacky way of passing page-specific data into the global `marked` extensions.