Adds tweet markdown plugin.
This renders markdown with the syntax:

```markdown

\`\`\`tweet
{
  // ...
}
\`\`\`

```

To use the provided JSON object as a configuration to render a tweet component inline in the document. The component mostly matches up to Twitter's UI, but with some simplifications.

I opted to use a hard-coded configuration here to avoid making real network requests to Twitter in order to render content either during SSG or CSR. This gives better performance (everything is prerendered), privacy (no network request to Twitter), and accessibility (it's all just HTML and CSS).

This uses Zod to validate the config schema and renders a Nunjucks template. I opted to keep the template in the same file to avoid making a filesystem read which would be asynchronous and fight against Marked. For the same reason, it also does not validate that the tweet actually exists. I'm hoping to make a tool in the future which will generate this configuration for a given tweet href.