Adds emoji support.
Most of the work here comes from following suggestions in this [excellent blog post](https://markoskon.com/creating-font-subsets/) which explains everything you ever wanted to know about font subsetting.

This uses the [Noto Color Emoji font](https://github.com/googlefonts/noto-emoji/) to display consistent emojis. Unfortunately there are a lot of emojis so the file is quite large (~10 MB). To give viable performance here, the font is subset to only the emojis that are actually used on the page (just one for testing right now). The `@font-face` rule also includes [`unicode-range`](https://developer.mozilla.org/en-US/docs/Web/CSS/@font-face/unicode-range) so the font is only downloaded if there is actually an emoji on the page. Even on slow 3G, this is still able to show in under 2 seconds.

The implementation is a bit unintuitive. Ideally, this would be done as part of the build process, but that is not easily achieveable because:
1. 11ty doesn't support an async main function yet, so it is not possible to read the filesystem and generate the font and its CSS file (at least without using sync FS hacks).
2. CSS files are bundled via the `css` pipe in Nunjucks, executed when the template for a page is rendered. Therefore we can't generate the file *after* 11ty runs, because it would be too late to bundle.
3. The subsetting tool (`pyftsubset`) is implemented in Python, so it is tricky to depend upon directly in a Node/NPM-based project.

However, the font only needs to be updated when a new emoji is used, which should happen infrequently. Instead the font is pre-built in source, with some documentation about how to update it when a new emoji is needed. This means running the Python tool is not a part of the build process and we don't need to worry about how to time things with 11ty.

I also took the liberty of updating fonts to be depended upon via the `theme.css` file where they are actually declared in the `font-family`. I think it's better to keep them in the CSS build graph rather than special casing it in a Nunjucks template.