Refactors emoji tooling to use a file with raw emoji text in it.
For some reason, the unicode code points were incompatible with the "beach with umbrella" emoji and `pyftsubset` would fail. Instead, using `--text-file` with emojis directly used works and is much more intuitive. The downside of this is that I can't include comments in the file, but the code points are in the `noto-color-emoji.css` file anyways and the README includes most of the documentation.

Also adds emoji for streamable HTML fragments blog post.