Adds spacing to adjacent paragraphs in blockquotes.
Two newlines in markdown blockquotes would only yield a single newline in the rendered HTML, which isn't really the intent and makes it appears like hard line breaks instead of line wrapped paragraphs. Adding the additional spacing makes it look like multiple paragraphs in the rendered output, which is more reflective of the intent.