Adds video markdown extension to better optimize video tags.
This sets the video size and supports two modes, demo videos and gifs. The former is a full screen video clip with controls shown for the user to pause or fullscreen for a better look. The latter is intended for a gif or meme clip, hides controls, and renders a little smaller.

I originally wanted to make the videos load lazily and require a `poster`. However both of these proved infeasible. Lazy loading is incompatible with autoplay videos, unless you require JavaScript for the video, which I am currently unwilling to do (https://techhub.social/@develwithoutacause/109702514595844465). An explicit `poster` just added another network request for an image which was difficult to optimize and didn't prevent the video from downloading eagerly either. A more in-depth approach which client-side renders a lazy-loaded video for JS-enabled browsers might improve this, but until then I'll stick with eagerly loading the videos.

Filed https://github.com/dgp1130/blog/issues/56 to follow up with optimizing videos.