Adds `<dwac-lazy />` component and uses it to lazy load videos.
Videos always seem to be loaded eagerly and have to `loading="lazy"` attribute or any other kind of deferred loading. While lazy loaded videos is possible, it requires JavaScript and is not directly `<noscript />` compatible. See: https://techhub.social/@develwithoutacause/109702514595844465. Instead, `<dwac-lazy />` makes a whole subcomponent tree render lazily, and supports `<video />` tags in the process.

`<dwac-lazy />` renders its content initially in a `<template />` and then clones and applies that content when it nears the viewport. This makes the `<video />` tag invisible to the browser until it is ready to be loaded. To avoid CLS, the `<dwac-lazy />` element is rendered with `aspect-ratio` so it reserves the correct amount of space in the initial layout, even before the `<video />` tag is rendered.

The component uses `IntersectionObserver` to wait until it intersects the viewport. A margin of `100%` is applied, so the video will start loading 1 viewport height before it is actually visible. This allows the video to preload a little bit without forcing all videos on the page to load eagerly.

It also requires a `<noscript />` tag, because the rendered `<template />` would be invisible to `<noscript />` users. In such a case, we're throwing away the performance optimizations, but it should be just as functional.