Adds social image.
This is specific resolution for Open Graph as suggested by https://www.opengraph.xyz/. It's also somewhat higher quality than a lot of the other images on this site because I don't have control over what size it will be displayed at, so I have to "round up" to a decently large size.

I'm only including `webp` because in a social link content, I don't exactly get a `<picture>` tag with multiple sources, so there's not much value in an AVIF version.