Adds `inline-html` extension.
This allows rendering raw HTML content from a markdown document like so:

```inline-html
<div>Hello, World!</div>
```

This will be rendered as is (not escaped and printed inside a `<code>` tag like ```html ... ``` would do). Marked already supports inline HTML so this `inline-html` code block is typically not necessary. However it preserves all internal whitespace in a way which is difficult/impossible to do for HTML inlined in a markdown document which has its own whitespace semantics. Therefore this is a more accurate representation of the provided content and is useful for examples where precise whitespace behavior is important.