Removes `nodegit`.
`nodegit` requires compilation from source with `nodegyp` which is painful during `npm install` and brittle to different development environments. For example, it requires `krb5-config` which is a thing I don't understand and doesn't appear to be installed by default on most systems.

Instead we subprocess the `git` command from the `$PATH` and parse `stdout`. This is programmatically more brittle because Git output isn't guaranteed to be stable and non-hermetic because Git's version is not pinned in the repository. However, Git doesn't change that frequently for this to be a major concern, and I see this as less brittle than depending on `nodegit`.