Adds RSS feed link to share widget.
This makes the RSS feed more discoverable for users. Technically the feed is not unique to the post they are reading, however most users will be linked directly to a single post and (if they like it) may consider adding the feed to their RSS reader without ever seeing the home page, so I think it's important to have this link on post pages somewhere.

I also adjusted the SVG icon slightly to "modernize" the look, getting rid of the rounded corners and gradient, instead going for a more minimalistic approach which feels like it fits the blog a bit better. I tweaked the view box as well to adjust the sizing so it more closely matches the other icons in the share widget.