Updates Twitter logo `<title>` to `<desc>`.
This `<title>` was overriding the `<a title="...">` value for the wrapping link, which meant that hovering over the icon just said "Twitter logo", not the action which will be taken on click "Share to Twitter". AIUI, `<desc>` is closer to `alt` text which is where "Twitter logo" would be more appropriate.