Upgrades `marked` to current latest.
Refs #26.

The main feature I want to take advantage of is `async` extensions to support `width` and `height` attributes on images and will be coming in a future commit. This update changed the parameter syntax for extensions which is the main thing which needed to be fixed. Those changes are technically optional, but the TypeScript types seem to require the new renderer API, so I'm not sure how else to avoid them.

The two most interesting changes are the the `highlight` API has been removed and heading IDs are not longer specified by default. I had to add two additional plugins to add those features back in, but they're relatively drop-in replacements. Unfortunately the new highlight extension doesn't seem to allow `return false` like it previously could, meaning there's no easy way to ignore unsupported languages. Other plugins use code block syntax, so ignoring languages is kind of important. Best I can do is put the highlight extension first and no-op the code. This seems to avoid wrapping in `<code>` tags or any other undesirable decoration.

I was also able to take advantage of the newly exposed `Marked` object, meaning I no longer need to rely on the singleton for tests. Each test gets its own complete environment which improves test isolation.