Fixes video size during initial load.
While `<dwac-lazy>` was being sized properly, the `<video>` element seems to default to a 2:1 aspect ratio. I'm a bit baffled as to how, Chrome's user agent styles default to `aspect-ratio: auto 1920 / 1080;`, but my attempts to fix it with something like `width: 100%; height: auto;` just haven't worked. The only solution I found was to copy the `aspect-ratio` style into the `<video>` tag as well to make sure it aligned with `<dwac-lazy>`. I feel like that shouldn't be necessary, but I can't find a better approach.