Fixes video bottom margin CLS.
There was some small CLS in the bottom margin of `<dwac-lazy>` because the styling was based on whether it contains a `<video>` tag. However, if the video enters the viewport immediately, the `<dwac-lazy>` may be visible while it still contains a `<template>` tag. Only once the JS loads does the `<template>` get converted into a `<video>` and the style takes effect. So instead we mark the `<dwac-lazy>` with `.video` so the style always applies, even before the template is stamped into the DOM.