Fixes `<dwac-share>` prompt.
This component was recently updated to a hard 44px height to avoid CLS, but if it renders a prompt string, the content overflows. To address this, I pulled the prompt string outside `<dwac-share>` so it can remain the hard-coded 44px regardless of the prompt being there. This also prerenders the prompt string which makes it a little more accessible.