Defines size adjusted Times New Roman as a fallback font for Lora.
This size adjustment brings Times New Roman roughly in line with Lora's sizing. This allows the browser to show this adjusted Times New Roman while downloading Lora. Once the download completes, the browser can swap it in with minimal CLS because the sizes match up. I generated the fallback font settings via the amazing [Fallback Font Generator](https://screenspan.net/fallback) tool.

I also switched the fallback font from Verdana to Times New Roman. I didn't realize that Verdana is actually sans-serif while Lora is serif. Aligning serif-ness makes the sizing more consistent and also just reduces visual jitter when the swap occurs.

Since the size is so close and the swap so minor, I decided to replace `font-display: fallback;` and `font-display: optional;` with `font-display: swap;`. We had both `fallback` and `optional` previously due to browser support. Browsers that supported `fallback` would use that by nature of it coming after `optional`. Other browsers would use `optional` since that's the only property they understand. `optional` has no swap period, meaning that if the font is not downloaded fast enough, the page will render without it to avoid blocking the user. `fallback` has a "short" swap period meaning it will replace the font after it is displayed to the user, but only for a short time. The idea is that we don't want the user to have a slow download of the font, start reading a post, and then have the page re-render with a different font while they're in the middle of reading it.

I decided on `font-display: swap;` instead which always replaces the font when its loaded. This is typically not recommended, but I think it is ok because the size adjust reduces visual noise from the swap. It should be relatively low impact to change the font, even if the user has already started reading.