Fixes spacing of share element.
The share element was in slightly different positions on the post page vs the index page because the `<main>` element had different padding. This aligns the padding of `<main>` and margin of `<dwac-share>` across the two pages. It also eliminates `.base-content` in favor of `main`. Since there should only be one `<main>` on the page, this should be safe.