Set image performance attributes for logo and author images.
The logo is above the fold, so I marked that with `loading="eager"` and `fetchpriority="high"` to ensure the browser prioritizes it. I'm a little worried browsers will prioritize it over the LCP image, which I don't want, but that should also have the same attributes and there's few enough resources loaded at the start that there shouldn't be too much network contention.

The author image is all the way at the bottom of the document, not just below the fold. I also tend to write long blog posts, so most users won't see this image at all. As a result, I gave that the laziest options of `loading="lazy" fetchpriority="low"`. I normally wouldn't go out of my way to specify `fetchpriority="low"`, but this seems like the image that most applies to.

I also added `decoding="async"` to both of them as I don't see any reason not to, regardless of how prominent the images are. The two other images I could find for opening a demo in a new tab and a Twitter avatar both got the generic `loading="lazy" decoding="async"`.