Use non-emoji fonts before emoji fonts.
Discovered today that when the fonts fail to load (either from network error, or blocked by extension), they will use the space from Noto Color Emoji. That font *does* include a space character, equivalent to one emoji width, which is quite large compared to spaces in other fonts. This makes the entire text fairly unreadable.

Based on [this comment](https://github.com/googlefonts/noto-emoji/issues/379#issuecomment-1030178518), it seems the recommendation is to place Noto *after* a non-emoji font. Since custom fonts can always fail to load, we can't rely on Lora to do this, so we need to follow `serif` to guarantee that a non-emoji font is used first. That font will claim the space character and use a normal width.

I also added an extra `emoji` font, so if Noto fails to load, hopefully the browser will fall back to some system font supporting the character. I'm not entirely sure that `serif` or `Times New Roman` is strictly guranteed *not* to support emojis. There is a risk that if one of those fonts includes an emoji, it will be preferred to Noto, which I would not want. However this seems to work as expected on a Windows and Linux test environment when blocking arbitrary custom fonts, so I *think* it's ok.

Two other weird data data points:
1. I'm only able to reproduce on Linux, Windows seems fine (both Chrome).
1. The bug only occurs if Lora *and* Noto fail to load. If Lora *xor* Noto fail, spacing is still normal. I get why Lora needs to fail, because normally it has priority to claim the space character. However, it seems particularly weird, because when Noto fails to load, the space character is attributed to Noto and given emoji-width. I don't follow why the browser would associate spaces with a font that failed to load (but not do so when the font succeeds - that's Lora failed, Noto succeed case), or how it would know how wide Noto expects emojis to even be (surely it would need the font file to know that?).

This seems very weird to me and I don't think I fully understand this bug, but I'm reasonably confident this is the correct solution, so YOLO?