Disable `collapseWhitespace`.
Comment elaborates on context, but the HTML whitespace post intentionally depends on not collapsing whitespacing and there's no easy way to opt-it out. I quick check of the page indicates the benefit is minimal anyways, so I'm inclined ot just turn it off.

If anything this just reinforces the point of that post that HTML is broken, given that my attempt at minification broke a blog post.