Narrow header styles to ones with real roles.
Demos for an upcoming blog post use `<h1>` and accidentally inherit this style, but I need to use `role="none"` anyways so they don't contribute confuse the semantic HTML, so this should also disable the link style.

There's probably better ways to do this automatically with some markdown plugin which opts-in markdown headers into this style with a class or something, but I'm going to keep it simple for now and re-evaluate if it becomes a persistent problem.