Remove `JSDom` dependency.
`JSDom` is not compatible with CSS `@layer` and fails to parse CSS containing those statements. We don't _really_ need `JSDom`, so instead I'm falling back to its `parse5` dependency which is able to make this work. The implementation is slightly more complicated as we can't use `document.querySelectorAll` and I also decided to pull in `@parse5/tools` primarily for the `getTextContent` function. But this seems good enough for the job and works with `@layer`.