Optimize `injectCsp`.
This refactors and optimizes `injectCsp` by extracting all the relevant resources in a single pass instead of iterating through the DOM for each resource kind. This also results in fewer overall generator functions and hopefully an easier to follow control flow. It also replaces the regex for `<head>` into leveraging the same single-pass parse to support `<head>` tags with attributes potentially containing `>` characters.

Ultimately this is probably a huge waste of time, the performance win isn't _that_ worth it and the odds of a `<head>` tag containing an attribute with a `>` is pretty minimal, however it satisfied a self-nerd-snipe.