Replace Karma with Vitest browser mode.
So I tried to upgrade to the next TS major version, but ran into weird Karma errors I could not understand and failed to find much resources on. I figured this was as good an opportunity as any and decided to migrate from Karma to Vitest browser mode. It's still technically experimental, but I figured it was worth trying out in a slightly more serious fashion.

I'll be honest, this was a lot harder than I thought. I ran into numerous technical issues which I cannot describe very well. A small sampling:
1. I now need to import Lit code with file extensions, which I did not before and don't know why.
2. I suddenly need `useDefineForClassFields: false` and don't understand why. Lit seems to be running into new errors it didn't with Karma. This could potentially be addressed by switching to standard decorators with a Babel transform, but I don't want to do that just yet. I also don't want to diverge the test execution order from the production execution order.
3. At multiple times I had weird install errors, either with Rollup not loading optional dependencies or suddenly generating CommonJS code. Reinstalling seemed to fix the issues eventually, but also did not fix it on multiple other occasions.
4. I need to manually call `vi.restoreAllMocks()` which I would expect to be done automatically.
5. I still don't like that type checking requires a different `tsc --noEmit` call which is awkward to work in and frequently hides errors when running in watch mode.
6. Suddenly TS is eliding non-value imports, which is an issue for the web component definitions which rely on top-level side effects. I had to manually use side-effectful imports which are otherwise already imported.
7. `@types/node` seems to be getting pulled in somehow for reasons I can't explain, confusing the return type of `setTimeout`.
8. The watch mode UI isn't great IMHO and seems to force you into a managed Playwright browser instance which I _really_ dislike. I much preferred Karma's approach of just printing a link and letting you navigate to it directly.
9. Lots more issues I can't even remember at this point...

Ultimately browser mode is still experimental, so weird issues are somewhat expected. I do anticipate it will be refined as Vitest progresses and this unblocks an immediate upgrade of TS, so it does solve my immediate problem and moves off deprecated Karma.