Skip builds in Netlify.
In an upcoming change to Vitest, Netlify gets confused between whether this is an Eleventy project or a Vite project and doesn't know what configuration to apply. Explicitly configuring the framework is enough to fix `netlify dev`, but still breaks deployments due to [a bug in Netlify](https://answers.netlify.com/t/netlify-ignores-framework-configuration-and-throws-error-detected-commands-for-eleventy-vite/156020).

This can be worked around with the [`--no-build`](https://cli.netlify.com/commands/deploy/) which I want to set anyways because I rely on pre-building the application before deployment, such that Netlify's build should not do anything. This should avoid accidentally relying on any framework integration I don't need.