Remove unimportant pages from search indexes.
These are "internal" pages not linked from the main site or intended for public consumption and should not be present in search indexes.

I initially tried to use `robots.txt` to disallow these pages, but Gemini states that this only prevents direct _crawling_ of the pages, and they might still be indexed based on other links, so a `<meta>` tag is a more robust solution.