Update Noto Color Emoji font.
Firefox was not accepting the emoji font due to a "no supported glyph data table(s) present". I tried the raw `noto-color-emoji-all.ttf` file and found the same error, so this seems like a bug in the font itself. I redownloaded a more up-to-date copy which does not seem to surface this issue, so I presume it has been fixed upstream. I re-ran the subsetting process and confirmed the error is still addressed, so updating the font is all that's necessary.

Subsetting did fail unless I installed `lxml`. Not sure exactly why this is necessary now, but maybe some new feature of the font requires this tool to process.